+++ src/components/explore/landcover/charts/landcover.js
